스펙트럼 추가, ugen bypass 추가. 근데 동기화 문제가 남아있다! 고쳐라!

+++ _library/PwElements/src/pw2/element/VUMeter.java
@@ -69,7 +69,7 @@ public class VUMeter extends Element {
   public static double linToDb(double in) {
     return 20 * Math.log10(in);
   }
-  public static double dbToLin(float in) {
+  public static double dbToLin(double in) {
     return Math.pow(10, 0.05 * in);
   }
   @Override
